My CI/CD tastes are eclectic and ever changing.
Turns out building the meta package with direct project references *does* get what I want: Built before the other packages are published, and I don't have to include the empty dll it produces, thanks to IncludeBuildOutput being false. I went ahead and renamed it to Meta.csproj, as that makes it seem less directly the "main" project, because it shouldn't be. It will still build the empty dll when built, but that's less annoying, so I've just got it skipped in the solution for now.

Removed nuget-meta.yaml, as I'm back to the single runner for now.