WinForms: Major reorganization:
Organized many classes into the folders (and namespaces) Components, Controls, Forms. Placed many more into an Extensions namespace/folder. This is going to involve a lot of breaking changes, but should help a lot with no longer needing to alias specific types that the extension classes were helping with. It was getting to the point of being cumbersome.

As well, the extension classes themselves have been renamed <theirname>Extensions.cs, since now a simple `using cmdwtf.Toolkit.WinForms.Extensions;` will get all of them and not clobber the local namespaces.